logged in user can change or delete ONLY it's own information

+++ routes/api.php
@@ -21,20 +21,22 @@ Route::get('/users/search/{name}', [AuthController::class, 'search']);
 // protected routes
 Route::group(['middleware' => ['auth:sanctum']], function () {
     // create new url list
-    Route::post('/users/{id}', [AuthController::class, 'url_post']);
+    Route::post('/users/{id}', [AuthController::class, 'url_post'])->middleware('auth', 'check_user_ownership');
     // change the informations of user by id
-    Route::put('/users/{id}', [AuthController::class, 'update']);
+    Route::put('/users/{id}', [AuthController::class, 'update'])->middleware('auth', 'check_user_ownership');
     // change the informations of url table by id
-    Route::put('/users/{id}/urls/{url_id}', [AuthController::class, 'url_update']);
-    // delete item from url table
-    Route::delete('/users/{id}/urls/{url_id}', [AuthController::class, 'url_destroy']);
+    Route::put('/users/{id}/urls/{url_id}', [AuthController::class, 'url_update'])->middleware('auth', 'check_user_ownership');
     // delete item from users table
-    Route::delete('/users/{id}', [AuthController::class, 'destroy']);
+    Route::delete('/users/{id}', [AuthController::class, 'destroy'])->middleware('auth', 'check_user_ownership');
+    // delete item from url table
+    Route::delete('/users/{id}/urls/{url_id}', [AuthController::class, 'url_destroy'])->middleware('auth', 'check_user_ownership');
     // user logout
-    Route::post('/logout', [AuthController::class, 'logout']);
+    Route::post('/logout', [AuthController::class, 'logout'])->middleware('auth', 'check_user_ownership');
     
     // // sent a new record to accounts table
     // Route::post('/accounts', [AccountController::class, 'store']);
+
+
 });
 
 Route::middleware('auth:api')->get('/user', function (Request $request) {
